padding-in-log-bug
